Basic filters for gtd, improved agenda display, added export to orgmode plugin, and started writing some unit tests.
GTD lists can now filter by context or TODO state via a form.

Agenda display now provides more information and is passed as a list instead of a queryset

the plugins/orgmode module now has export_to_string that takes a node and returns the output string that
could be written to a file. It has not passed unit testing yet (the schedule element needs work).

Unit tests were added to examine the orgmode plugin. It checks input against output. The test string is not
terribly complex yet, this should be changed in the future.