Basic inbox functionality, Color object and activated capture widget.
Wolfmail app now exists, currently only supports the inbox view. A MailItem
class defines a piece of mail and a Label class defines an IMAP-like folder.

A label can have a color associated with it, so I had to make a Color object in orgwolf.
It holds RGB and alpha information for processing by various methods.

The capture widget that appears on each page now actually adds a new MailItem with the inbox
label for later processing.