Re-wrote org-mode plugin without using PyOrgMode.
PyOrgMode was ok but was a somewhat klunky way to import org-mode files.
Basically the file had to be parsed and the parsed data then had to be reparsed.
I re-wrote the import_structure() method to handle this job natively. Unittests were
also written to test the required regular expressions. Some small changes were made to
GTD models. The stacks.py file implements a basic stack functionality (which is needed
for keeping track of which level we're on while walking through project nodes).