Reimplemented node closing timestamps as a pre_save signal.
Previously, editing a node via the projects.views.edit_node or
projects.views.new_node views would automatically set the Node.closed
timestamp before saving. This functionality was implemented as a seperate
function that catches the pre_save signal for the Node class. This
means that the timestamp will be set no matter how the todo_state is
changed (including the admin interface).