Reworked what a project is.
Previously, a `class Project` was defined. During creation of a top-level `Node` object,
a new `Project` object was created that matched the new `Node`. This led to potential
descynchronization.

Now, a project is defined as a `Node` with no `parent` attribute.
`Node.related_name = ManyToManyField` holds a list of other top-level nodes related to this Node.
Functionality has been coded into `NodeForm.__init__()` to populate the widget and
`NodeForm.clean_related_fields()` enforces the constraints of only using top-level nodes.
The admin page for the `Node` object also references the `NodeForm` class so it takes
advantage of this validation.