Preliminary collapsing node functionality and ajax framework
Added some basic functionality for collapsing nodes. This required adding in
the ability to set ajax endpoints. Right now this uses the django-ajax module
but may be rewritten natively in the future. Added a framework for QUnit tests
as well. Some refactoring of CSS and javascript moved code from base.html template
to their own files for more atomic testing of javascript.